Changed source number field to double; made source number searchable by parent source

+++ src/main/java/com/toohuman/controller/source/SourcesKeywordSearchController.java
@@ -63,7 +63,7 @@ public class SourcesKeywordSearchController {
 		}
 		workingSet.addAll(repo.findByCollection(keyword));
 		try {
-			workingSet.addAll(repo.findBySourceNumber(Integer.parseInt(keyword)));
+			workingSet.addAll(repo.findBySourceNumber(Double.parseDouble(keyword)));
 		} catch(Exception e) {
 //			System.out.println("NaN entered as sourceNumber");
 		}
